docs: add missing docs
