+++ Sources/DesignTokensGenerator/Extensions/URL+Extension.swift
@@ -1,6 +1,7 @@
 import Foundation
 
 extension URL {
+  /// A `Bool` indicating whether the `URL` represents a directory.
   var isDirectory: Bool {
     let values = try? resourceValues(forKeys: [.isDirectoryKey])
     return values?.isDirectory ?? false
