+++ Sources/DesignTokensGenerator/Generators/SourceCodeGenerator.swift
@@ -2,6 +2,7 @@ import Foundation
 import Stencil
 import StencilSwiftKit
 
+/// A protocol defining requirements for a source code generator.
 protocol SourceCodeGenerator {
   func generate(with environment: Stencil.Environment) throws -> [SourceCodeFile]
 }
