+++ Sources/DesignTokensGenerator/SourceCodeFile.swift
@@ -1,7 +1,13 @@
 import Foundation
 
+/// A type representing a source code file.
 struct SourceCodeFile {
+  
+  // MARK: - Stored Properties
+  
+  /// The name of the file.
   let name: String
 
+  /// The content of the file.
   let content: String
 }
